Completed search functionality and movie description page #2
