+++ client/package.json
@@ -15,6 +15,7 @@
     "react-icons": "^4.8.0",
     "react-router-dom": "^6.10.0",
     "react-scripts": "5.0.1",
+    "react-spinners": "^0.13.8",
     "reacticon": "^0.0.1",
     "web-vitals": "^2.1.4"
   },
