+++ client/src/Home.css
@@ -91,6 +91,10 @@
     opacity: 0.6;
  }
 
+ .movie-card:hover {
+  background-color: #7f99e1;
+ }
+
 
 @media only screen and (min-width: 784px) {
   .film-icon {
