+++ client/src/SearchResults.css
@@ -1,0 +1,19 @@
+.search-results,
+.movie-description {
+  background-color: #111;
+  height: 100vh;
+}
+
+.rating-border {
+  border: 1px solid transparent;
+  border-radius: 0.3em;
+  background-color: #7f99e1;
+}
+
+.rating {
+  margin-left: 0.5rem;
+}
+
+@media only screen and (min-width: 784px) {
+
+}
